requirements.txt
streamlit>=1.35.0
ultralytics==8.2.0
torch==2.5.1
torchvision==0.20.1
opencv-python==4.9.0.80
numpy==1.26.4
pandas==2.2.2
detectron2@git+https://github.com/facebookresearch/detectron2.git@v0.6
Pillow==9.5.0
pycocotools==2.0.7